% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{findASevents}
\alias{findASevents}
\title{Classify alternative splicing events and its splice junctions}
\usage{
findASevents(gtf, min_RI_length = 10)
}
\arguments{
\item{gtf}{Can be path to a GTF transcriptome file or a GenomicRanges object
containing exon and transcript coordinates of each transcript. The latter
object should contain the metadata gene_id, transcript_id and type.}

\item{min_RI_length}{Minimum retained intron length}
}
\description{
Classify alternative splicing events and its splice junctions
}
