% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test.gtf}
\alias{test.gtf}
\title{Mock GenomicRanges object in GTF format}
\format{
A GRanges object with 35 ranges and 4 metadata columns:
\describe{
  \item{ranges}{Chromosome, start, end, and strand info of 4 transcripts
  and its exons}
  \item{type}{Entry type; transcript or exon}
  \item{transcript_id}{Name or ID given to transcripts}
  \item{gene_id}{Name or ID given to gene origin of transcripts}
  ...
}
}
\usage{
data(test.gtf)
}
\description{
An object cotaining for 4 hypothetical transcripts from "geneA" with different
exon structures
}
\keyword{datasets}
